requirements.txt
gradio
tensorflow
pillow
numpy
scikit-learn